ai-worker/requirements.txt
demucs==4.0.1
openai-whisper==20231117
crepe==0.0.16
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
librosa>=0.10.0
pika>=1.3.0
redis>=5.0.0
boto3>=1.28.0
python-dotenv>=1.0.0
pydub>=0.25.0
soundfile>=0.12.0
